lot-as-component
переделал лоты под компоненту
