+++ client/src/components/Lot/Lot.scss
@@ -1,0 +1,6 @@
+.market-lot {
+    border: 3px solid #000000;
+    margin: 10px;
+    border-radius: 5px;
+    padding: 10px;
+}
