+++ client/src/components/MarketTab/MarketTab.scss
@@ -1,11 +1,4 @@
 
-.market-lot {
-    border: 3px solid #000000;
-    margin: 10px;
-    border-radius: 5px;
-    padding: 10px;
-}
-
 .lots {
     overflow-y: scroll;
     max-height: 400px;
